fix review5: GreenbidsInvocationService + to bean
